Split resolve module
